refactor(money): standardize Money handlig with BigDecimal and general cleanup

+++ bankback/src/main/java/com/app/bankback/controller/AccountsController.java
@@ -32,7 +32,7 @@ public class AccountsController {
         ops.transfer(
                 request.getFromAccountId(),
                 request.getToAccountId(),
-                Money.of(request.getAmount().doubleValue()),
+                Money.of(request.getAmount()),
                 request.getRequestOwnerId()
         );
         return ResponseEntity.noContent().build();
